backend/requirements.txt
fastapi==0.115.6
langchain==0.3.13
passlib==1.7.4
protobuf==5.29.2
pydantic==2.10.4
PyJWT==2.10.1
pymongo==4.10.1
